feat: added AddBiochemicalReactions method to model builder; refactored metabolomics_example.h to move core components to data simulator
